Final update
Implemented code for symbols and numbers toggles: split the characters array in three different arrays each one realted to a specific type of unit.

Code refactoring with implementation of function to generate Password with options of numbers and symbols.

Implemented classList DOM for dark colour when the toggle for dark/light mode is active. There are some conflicts in CSS that do not allow to manipulate .label as a class in order to add dark colour to the span for symbols, numbers, and password length. Styling is out of scope for this project, thus I opted to leave it as it is.